+++ SpringBoot-api/src/main/java/com/springBootApi/SpringBootapi/repository/FlightDataRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FlightDataRepository extends JpaRepository<FlightDatas, Integer> {
+public interface FlightDataRepository extends JpaRepository<FlightDatas, Long> {
 
 }
